// src/app/core/models/Employee.model.ts
export class  Employee{
    id: number = 0;
    tipoIdentificacion: string = "";
    identificacion: number;
    nombres: string = "";
    apellidos: string = "";
    telefono: number;
    email: string= "";
    edad: number = 0;
    sexo: string = "";
    fechaNacimiento: string = "";
    ocupacion: string = "";
    direccion: string = "";
    rol: string = "";
    estado: string = "";
    password: string = "";
  }
  